+++ CHANGELOG.md
@@ -26,6 +26,15 @@
 4.8.4 (unreleased)
 --------------------------
 
+### Improvements
+
+- [dev] Update the build environment to Xcode 15.4.
+
+
+### Fixes
+
+- Fix a typo in the Czech localization.
+
 
 
 4.8.3 (649)