+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 
 ### Improvements
 
+- Enable toolbar/menu items for the document editor even when the focus in on the inspector.
 - Display a warning if custom setting names contain “:”.
 - [dev] Update the build environment to Xcode 15.4.