+++ CHANGELOG.md
@@ -7,13 +7,14 @@
 
 - Revert the “Do Nothing” option for the “When nothing else is open” setting
 - Enable toolbar/menu items for the document editor even when the focus in on the inspector.
-- Display a warning if custom setting names contain “:”.
+- Display a warning on setting renaming if the name contains “:”.
 - [dev] Update the build environment to Xcode 15.4.
 
 
 ### Fixes
 
 - Fix an issue on CotEditor 4.8.3 that The Open dialog always shows up on launch despite the setting when the “Reopen windows from last session” option is disabled.
+- Fix an issue on the version browsing that document windows could become occasionally transparent.
 - Fix a typo in the Czech localization.